test(fixtures): centralize gated service setup
